chore: move test "helpers" to "testUtils
