+++ test/relations/one-to-one.operations.test.ts
@@ -1,5 +1,5 @@
 import { nullable, oneOf, primaryKey } from '../../src'
-import { testFactory } from '../helpers'
+import { testFactory } from '../testUtils'
 
 /**
  * Non-nullable one-to-one relationship.
